// packages/vgraph/src/typings/graph.ts
export type AnimateConfigs = {
  /**
   * The target entity that the animation applied to.
   * 动画作用的图中元素
   */
  target?: Node | Edge;
  /**
   * The type of animation.
   * 动画类型
   */
  type?: string;
  /**
   * Customized animation configurations, often related to animation type.
   * 动画个性化配置，往往与类型相关
   */
  custom?: {
    x?: number;
    y?: number;
    r?: number;
    size?: number[];
    color?: string;
    opacity?: number;
    [k: string]: any;
  };
  /**
   * General animation configurations.
   * 动画通用配置
   */
  common?: {
    /**
     * The duration of animation.
     * 动画时长
     */
    duration?: number;
    /**
     * The target shape that the animation applied to.
     * 动画作用的图形
     */
    target?: Shape;
    /**
     * Specify id of the animation.
     * 指定动画 id
     */
    id?: string;
    /**
     * Configurations involved in the animation.
     * 参与动画的属性
     */
    configs?: Record<string, any>;
    /**
     * A function called each frame.
     * 每帧的回调函数
     */
    onFrame?: (ratio: number) => any;
    /**
     * Whether the animation repeats.
     * 动画是否重复
     */
    repeat?: boolean | number;
    /**
     * Easing function of the animation.
     * 动画的缓动函数
     */
    easing?:
      | "easelinear"
      | "easeCubic"
      | "easePoly"
      | "easeQuad"
      | "easeSin"
      | "easeExp"
      | "easeBounce";
    /**
     * Delay execution time of the animation.
     * 动画延迟执行时间
     */
    delay?: number;
    /**
     * A function called after the animation is finished.
     * 动画执行结束回调
     */
    onFinish?: () => void;
  };
};